diff data no longer shown for added/deleted + Sakila3 tests
In the case of sources/columns/relationships that are added or deleted,
the "diff" data is no longer shown because it is redundant and noisy. If
the user wants to see the data for an added source, they can just look
at the added source themselves (also, sugar methods may still be added
to return this info, but it is the same for those cases -- it can just
query the old/new objects directly).

Also made a number of random types of changes to the Sakila3 test schema
sources to test detection of different cases of changes -
add/delete/change for sources, columns and relationships are all
now represented in test cases...