PRELIMINARY WORKING WILDCARD/GLOB FILTERING
Using the features added as of the dev version Hash::Layout 1.02_01
filtering out sources, relationships, columns etc can now be done with
wild card matching. For instance:

 ->filter_out('FOO_')

would exclude all surce names which start with 'FOO_' whether it is 1 or
100. The same works in other parts of the composite path like:

 ->filter_out('*is_nullable')
 ->filter_out('*:relationships/prd*')

The above would exclude all attr changes to 'is_nullable' and all
relationships in all tables which start with 'prd'

Also, reverted the changes of the previous commit which started to
establish a 6th type 'db_schema' which may or may not be needed yet
after all. The biggest blocker that this feature would have solved is
actually mostly solvable by simply coordinating/matching the 'db_schema'
and 'qualify_objects' schema loader options (for the case of comparing
dbic schema with direct DDL). Simply must incoude the db_schema prefix
in the table name on both sides. This still doesn't solve the conceptual
case when these cannot be matched, however, it may be better to add
support for 'table_name' in the ->filter_out arguments and now take
advantage of the string pattern matching, to essentiually ignore tjose
sort of chnages, however, this may not be feasible as changes in table
names will always show up ad delete/add, as the table it the top-level
thing. This just needs some more thought, for now we are sticking with
the same 5 type API

Set dev version 1.07_01