FIXES TO MANY WILDCARD MATCH SCENARIOS + UNIT TESTS
Lots of improvements and handling cases to work correctly with the new
wildcard filter matching. Added a bunch of new unit tests which test
this functionality.

The one exception which does not yet work as expected is 'isa' because
it is very different from the rest of the achema data structure. It is
an arrayref of strings (SOMETIMES) with +/-. If I could go back in time,
I would have not used this structure but it is too late for that which
means this must be supported. Getting it to work will involve jumping
through a lot of hoops ...

Requires Hash::Layout 1.02_04 with a critical fix to the glob matching
with '/' in values.

Set dev version 1.07_03