Add support for "diffing" on only the new_schema
The 'old_schema' peram is now optional, and if ommited, the diff will
return 'changed' events as if every single data item was modified. The
events report as 'changed' not 'added' since the added mode does not
provide a diff packet. This is essentially another way to view the
Schema data (SchemaData) but in the structure of a Diff. This is being
added so that the ->filter methods can be used for the same function as
the ->prune methods currently work on the SchemaData object. The
SchemaData ->prune API is a very limted version of the ->filter API
which is now even more significant with the wildcard matching added in
v1.10. By adding this option, the ->prune API becomes superfluous and
there is no need to enhance it, which is not actually very simple since
the SchemaData data structure isn't the same as the diff structure.

The purpose here is to start doing checksum versions by running a
checksum on the diff packet instead of the SchemaData packet, and
abandon the 'schemsum' API that currently resides in SchemaData

IN PROGRESS...

(set dev version v1.10_01)