Separate out single schema diff into its own class + fingerprint
For a more sane API, the logic path that allows a diff on new_schema
only has been moved into its own class State to establish that the API
is different and also in order to be able to test the Diff object later
on to see which it is. It should be a "class" taxonomy-wise and now it
is.

Also added fingerprint meythod to Diff

IN PROGRESS...

(set dev version 1.10_02)