References to parent objects should be weakened
The tree of objects from Schema to Source and from Source to InfoPacket
have back references from the child to the parent.  If not weakened,
these create a circular reference and memory leak.  It is unlikely that
the diffs would be produced very many times in a long-running program,
but better to plug the leak regardless.